# lab4/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  serializer.CurrentJSON:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      current_id:
        type: integer
      finish_date:
        type: string
      form_date:
        type: string
      moderator_login:
        type: string
      status:
        type: string
      voltage_bord:
        type: number
    type: object
  serializer.DeviceJSON:
    properties:
      coeff_efficiency:
        type: number
      coeff_reserve:
        type: number
      current_required:
        type: number
      description:
        type: string
      device_id:
        type: integer
      image:
        type: string
      in_stock:
        type: boolean
      is_delete:
        type: boolean
      name:
        type: string
      power_nominal:
        type: number
      resistance:
        type: number
      type:
        type: string
      voltage_nominal:
        type: number
    type: object
...
host: localhost
info:
  contact:
    email: support@current.com
    name: API Support
    url: http://localhost
  description: API для определения необходимой силы тока
  license:
    name: MIT
  title: Current API
  version: "1.0"
paths:
...
  /devices:
    get:
      description: Возвращает все устройства или фильтрует по названию
      parameters:
      - description: Название устройства для поиска
        in: query
        name: device_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список устройств
          schema:
            items:
              $ref: '#/definitions/serializer.DeviceJSON'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список устройств
      tags:
      - devices
  /devices/{id}:
    get:
      description: Возвращает информацию об устройстве по её идентификатору
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные устройства
          schema:
            $ref: '#/definitions/serializer.DeviceJSON'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить устройство по ID
      tags:
      - devices
  /devices/{id}/add-to-current-calculation:
    post:
      description: Добавляет устройство в заявку-черновик пользователя
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Расчёт с добавленным устройством
          schema:
            $ref: '#/definitions/serializer.CurrentJSON'
        "201":
          description: Создан новый расчёт
          schema:
            $ref: '#/definitions/serializer.CurrentJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Устройство уже в расчёте
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить устройство в расчёт
      tags:
      - devices
  /devices/{id}/delete-device:
    delete:
      description: Выполняет логическое удаление устройство по ID
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус удаления
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить устройство
      tags:
      - devices
  /devices/{id}/edit-device:
    put:
      consumes:
      - application/json
      description: Обновляет информацию об устройстве по ID
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные устройства
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/serializer.DeviceJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленное устройство
          schema:
            $ref: '#/definitions/serializer.DeviceJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить данные устройства
      tags:
      - devices
  /devices/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для устройства и возвращает обновленные данные
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение устройства
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Статус загрузки и данные устройства
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос или файл
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение устройства
      tags:
      - devices
  /devices/create-device:
    post:
      consumes:
      - application/json
      description: Создает новое устройство и возвращает его данные
      parameters:
      - description: Данные нового устройства
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/serializer.DeviceJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Созданное устройство
          schema:
            $ref: '#/definitions/serializer.DeviceJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новое устройство
      tags:
      - devices
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"